move _set_cv sklearn version to base
