+++ mne/decoding/time_gen.py
@@ -8,6 +8,7 @@
 import numpy as np
 import copy
 
+from .base import _set_cv
 from ..io.pick import _pick_data_channels
 from ..viz.decoding import plot_gat_matrix, plot_gat_times
 from ..parallel import parallel_func, check_n_jobs
@@ -1558,37 +1559,3 @@ def _chunk_data(X, slices):
     slices_chunk = [sl - start for sl in slices]
     X_chunk = X[:, :, start:stop]
     return X_chunk, slices_chunk
-
-
-def _set_cv(cv, clf=None, X=None, y=None):
-    from sklearn.base import is_classifier
-
-    # Set the default cross-validation depending on whether clf is classifier
-    # or regressor.
-    if check_version('sklearn', '0.18'):
-        from sklearn.model_selection import (check_cv, StratifiedKFold, KFold)
-        if isinstance(cv, (int, np.int)):
-            XFold = StratifiedKFold if is_classifier(clf) else KFold
-            cv = XFold(n_folds=cv)
-        cv = check_cv(cv=cv, y=y, classifier=is_classifier(clf))
-    else:
-        from sklearn.cross_validation import (check_cv, StratifiedKFold, KFold)
-        if isinstance(cv, (int, np.int)):
-            if is_classifier(clf):
-                cv = StratifiedKFold(y=y, n_folds=cv)
-            else:
-                cv = KFold(n=len(y), n_folds=cv)
-        cv = check_cv(cv=cv, X=X, y=y, classifier=is_classifier(clf))
-
-    # Extract train and test set to retrieve them at predict time
-    if hasattr(cv, 'split'):
-        cv_splits = [(train, test) for train, test in
-                     cv.split(X=np.zeros_like(y), y=y)]
-    else:
-        # XXX support sklearn.cross_validation cv
-        cv_splits = [(train, test) for train, test in cv]
-
-    if not np.all([len(train) for train, _ in cv_splits]):
-        raise ValueError('Some folds do not have any train epochs.')
-
-    return cv, cv_splits
